# api/swagger.yaml
openapi: 3.0.3
info:
  title: Auth Service API
  description: |
    API для сервиса аутентификации.
    
    Сервис предоставляет API для аутентификации и авторизации пользователей.
    
    ## Особенности
    
    - Получение пары токенов (access и refresh) для пользователя с указанным идентификатором
    - Обновление пары токенов
    - Получение GUID текущего пользователя (защищенный маршрут)
    - Деавторизация пользователя
    
    ## Требования к токенам
    
    ### Access токен
    - Формат: JWT
    - Алгоритм подписи: SHA512
    - Не хранится в базе данных
    
    ### Refresh токен
    - Произвольный формат
    - Передается в формате base64
    - Хранится в базе в виде bcrypt хеша
    - Защищен от повторного использования
    - Защищен от изменений на стороне клиента
    
    ## Особенности операции refresh
    - Операцию можно выполнить только той парой токенов, которая была выдана вместе
    - При изменении User-Agent операция запрещается и пользователь деавторизуется
    - При попытке обновления с нового IP отправляется POST-запрос на заданный webhook
  version: 1.0.0
  contact:
    name: API Support
    email: support@example.com
servers:
  - url: http://localhost:8080
    description: Локальный сервер разработки
  - url: https://api.example.com
    description: Продакшн сервер
...
paths:
...
  /auth/refresh:
    post:
      tags:
        - auth
      summary: Обновление пары токенов
      description: Обновление пары токенов с использованием refresh токена
      operationId: refreshTokens
      requestBody:
        description: Refresh токен для обновления
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshRequest'
      responses:
        '200':
          description: Успешное обновление токенов
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenPair'
        '400':
          description: Неверный формат запроса
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Неверный refresh токен или несоответствие User-Agent
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  schemas:
    TokenPair:
      type: object
      properties:
        access_token:
          type: string
          description: JWT токен доступа
          example: eyJhbGciOiJIUzUxMiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIzNWZlMTM1OC1hMDNhLTRiYmEtYjJhYS1iMGIxYzQ5MjY0ZDIiLCJleHAiOjE3MTY0NzY4MDB9.ACCESS_TOKEN
        refresh_token:
          type: string
          description: Refresh токен в формате base64
          example: ZXlKaGJHY2lPaUpJVXpVeE1pSXNJblI1Y0NJNklrcFhWQ0o5LmV5SnpkV0lpT2lJek5XWmxNVE0xT0MxaE1ETmhMVFJpWW1FdFlqSmhZUzFpTUdJeFl6UTVNalkwWkRJaUxDSmxlSEFpT2pFM01UWTBPRGM0TURCOUxnLllPVVJfUkVGUkVTSF9UT0tFTg==
      required:
        - access_token
        - refresh_token
    RefreshRequest:
      type: object
      properties:
        refresh_token:
          type: string
          description: Refresh токен в формате base64
          example: ZXlKaGJHY2lPaUpJVXpVeE1pSXNJblI1Y0NJNklrcFhWQ0o5LmV5SnpkV0lpT2lJek5XWmxNVE0xT0MxaE1ETmhMVFJpWW1FdFlqSmhZUzFpTUdJeFl6UTVNalkwWkRJaUxDSmxlSEFpT2pFM01UWTBPRGM0TURCOUxnLllPVVJfUkVGUkVTSF9UT0tFTg==
      required:
        - refresh_token
...
    Error:
      type: object
      properties:
        message:
          type: string
          description: Сообщение об ошибке
          example: Invalid token
      required:
        - message
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT токен доступа в формате Bearer